fix: make REPL auth-agnostic + Auth0 SPA login for playground (#14)
* chore: replace playground shell-loop with shared ReplSession

- Remove custom configure/mount/flush handlers from playground
- Use ReplSession from @tigrisdata/agent-shell/repl
- Implement ReplIO for xterm.js with prompt support
- Remove credentials.ts (session manages state)
- Update welcome message with login and configure options


* fix: move device flow to cli, make REPL auth-agnostic

- src/repl/auth.ts: slimmed to types only (LoginFn, LoginResult, Organization)
- src/cli/auth.ts: device flow implementation moved here (CLI-specific)
- ReplSession accepts loginFn via constructor — no hardcoded auth
- CLI passes deviceLogin, playground passes browserLogin


* chore: add Auth0 SPA login for playground, fix prompt redraw

- playground/src/auth.ts: browser login via @auth0/auth0-spa-js
- ShellLoop passes browserLogin to ReplSession
- Fix prompt redraw showing $ instead of "Enter number: " during org selection


---------
