+++ packages/agent-shell/playground/package.json
@@ -8,7 +8,7 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
-		"@tigrisdata/agent-shell": "^0.1.3",
+		"@tigrisdata/agent-shell": "^0.3.0",
 		"@xterm/addon-fit": "^0.10.0",
 		"@xterm/xterm": "^5.5.0"
 	},
