+++ packages/agent-shell/src/repl/index.ts
@@ -1,4 +1,4 @@
-export type { LoginResult, Organization } from "./auth.js";
-export { deviceLogin } from "./auth.js";
+export type { LoginFn, LoginResult, Organization } from "./auth.js";
 export type { ReplIO } from "./io.js";
+export type { ReplSessionOptions } from "./session.js";
 export { ReplSession } from "./session.js";
